+++ src/bump_vec.rs
@@ -1,11 +1,10 @@
 mod into_iter;
 
 use crate::{
-    bump_down, error_behavior_generic_methods_allocation_failure,
+    bump_down, error_behavior_generic_methods_allocation_failure, owned_slice,
     polyfill::{nonnull, pointer, slice},
-    up_align_usize_unchecked, BaseAllocator, BumpBox, BumpScope, Drain, ErrorBehavior, ExtractIf, FixedBumpVec,
-    GuaranteedAllocatedStats, MinimumAlignment, NoDrop, SetLenOnDropByPtr, SizedTypeProperties, Stats,
-    SupportedMinimumAlignment,
+    up_align_usize_unchecked, BaseAllocator, BumpBox, BumpScope, ErrorBehavior, FixedBumpVec, GuaranteedAllocatedStats,
+    MinimumAlignment, NoDrop, SetLenOnDropByPtr, SizedTypeProperties, Stats, SupportedMinimumAlignment,
 };
 use allocator_api2::alloc::Allocator;
 use core::{
@@ -1431,7 +1430,7 @@ where
     /// v.drain(..);
     /// assert_eq!(v, []);
     /// ```
-    pub fn drain<R>(&mut self, range: R) -> Drain<'_, T>
+    pub fn drain<R>(&mut self, range: R) -> owned_slice::Drain<'_, T>
     where
         R: RangeBounds<usize>,
     {
@@ -1491,7 +1490,7 @@ where
     /// ```
     ///
     /// [`retain`]: Self::retain
-    pub fn extract_if<F>(&mut self, filter: F) -> ExtractIf<T, F>
+    pub fn extract_if<F>(&mut self, filter: F) -> owned_slice::ExtractIf<T, F>
     where
         F: FnMut(&mut T) -> bool,
     {