+++ src/owned_slice.rs
@@ -1,0 +1,7 @@
+mod drain;
+mod extract_if;
+mod into_iter;
+
+pub use drain::Drain;
+pub use extract_if::ExtractIf;
+pub use into_iter::IntoIter;